+++ invoke/runners.py
@@ -1136,6 +1136,13 @@ class Result(object):
 """.format(x, val.rstrip()) if val else u"(no {0})".format(x))
         return u"\n".join(ret)
 
+    def __repr__(self):
+        # TODO: more? e.g. len of stdout/err? (how to represent cleanly in a
+        # 'x=y' format like this? e.g. '4b' is ambiguous as to what it
+        # represents
+        template = "<Result cmd={!r} exited={}>"
+        return template.format(self.command, self.exited)
+
     @property
     def ok(self):
         """