+++ sites/www/changelog.rst
@@ -2,6 +2,8 @@
 Changelog
 =========
 
+* :feature:`-` `Result <invoke.runners.Result>` objects now have a more useful
+  ``repr()``.
 * :bug:`432` Tighten application of IO thread ``join`` timeouts (in `run
   <invoke.runners.Runner.run>`) to only happen when :issue:`351` appears
   actually present. Otherwise, slow/overworked IO threads had a chance of being